Split ParcelDetails into separate component
