+++ frontend/src/features/mapSideBar/components/tabs/HeaderDivider.tsx
@@ -1,0 +1,6 @@
+import { Divider } from '@mui/material';
+import React from 'react';
+
+export const HeaderDivider = () => (
+  <Divider sx={{ mt: '5px', mb: '10px', height: '1px', background: '#1a57c7', opacity: '100%' }} />
+);
